fix(animation): ACT-1005 useAnimations hook
